backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic[email]==2.5.3
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.6
email-validator==2.1.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
slowapi==0.1.9
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.4
